thales/backend/requirements.txt
aiohttp==3.7.3
aioredis==1.3.1
aniso8601==8.0.0
apispec==4.0.0
asgiref==3.3.1
astroid==2.3.2
async-timeout==3.0.1
attrs==20.2.0
autobahn==21.2.1
Automat==20.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.3.2
boto3==1.14.59
botocore==1.17.59
cached-property==1.5.2
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.1
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
daphne==3.0.1
decorator==4.4.2
distro==1.5.0
Django==3.1.3
django-channels==0.7.0
djangorestframework==3.12.2
docker==4.4.4
docker-compose==1.28.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
docx==0.2.4
docx2python==1.26
fastparquet==0.4.1
feedfinder2==0.0.4
feedparser==5.1.3
filelock==3.0.12
Flask==0.10.1
Flask-Cache==0.13.1
Flask-RESTful==0.3.8
flask-restplus==0.13.0
forex-python==1.5
future==0.15.2
gevent==21.1.2
google==2.0.3
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-datastore==1.10.0
google-pasta==0.2.0
googleapis-common-protos==1.51.0
GoogleNews==1.4.1
greenlet==1.0.0
grip==4.5.2
grpcio==1.27.2
gunicorn==19.4.5
h5py==2.10.0
hiredis==1.1.0
httplib2==0.17.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==1.7.0
incremental==17.5.0
install==1.3.4
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==0.24
jedi==0.17.2
jieba3k==0.35.1
Jinja2==2.8
jmespath==0.9.5
joblib==0.14.0
jsonlines==1.2.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
Keras==2.4.3
lazy-object-proxy==1.4.2
llvmlite==0.34.0
lxml==4.5.2
Markdown==3.3.3
MarkupSafe==0.23
mccabe==0.6.1
msgpack==1.0.2
multidict==5.1.0
myapp==0.1.dev0
newspaper3k==0.2.8
nltk==3.4.5
nose==1.3.7
numba==0.51.2
numpy==1.17.3
oauthlib==3.1.0
packaging==20.4
pandas==0.25.2
paramiko==2.7.2
parso==0.7.1
path-and-address==2.0.1
pickleshare==0.7.5
Pillow==7.1.2
pipreqs==0.4.10
prompt-toolkit==3.0.7
protobuf==3.11.3
protobuf3-to-dict==0.1.5
psutil==5.7.0
psycopg2==2.8.6
pyandoc==0.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.1
PyHamcrest==2.0.2
pylint==2.4.3
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.0
python-docx==0.8.10
python-dotenv==0.15.0
pytz==2019.3
pywin32==227
PyYAML==5.3.1
pyzmq==19.0.2
quora==0.1.22
redis==3.5.3
regex==2020.6.8
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.43
sagemaker==2.5.5
scikit-learn==0.23.1
scipy==1.5.1
sentencepiece==0.1.91
seqeval==0.0.12
service-identity==18.1.0
sgmllib3k==1.0.0
simplejson==3.17.2
simpletransformers==0.45.0
six==1.12.0
slackclient==2.9.3
slacker==0.9.65
slacker-log-handler==1.8.0
smdebug-rulesconfig==0.1.5
soupsieve==1.9.5
SQLAlchemy==1.3.20
sqlparse==0.4.1
tensorboardX==2.1
textblob==0.15.3
texttable==1.6.3
threadpoolctl==2.1.0
thrift==0.13.0
tinysegmenter==0.3
tldextract==2.2.3
tokenizers==0.8.1rc2
tornado==6.0.4
tqdm==4.48.0
traitlets==5.0.4
transformers==3.0.2
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==16.7.6
wcwidth==0.2.5
websocket==0.2.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
wrapt==1.11.2
xlrd==1.2.0
yarg==0.1.9
yarl==1.6.3
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.2.0
